+++ src/features/paper-list/models/paper-list.constants.ts
@@ -1,5 +1,7 @@
 export const RECIPIENTS_LIMIT = 5
+
 export const RECIPIENTS_OFFSET = 0
+
 export const RECIPIENTS_TYPE = {
   RECENT: "recent",
   POPULAR: "popular"
